refactor : divide knowledge apis #48
